Centralize media action queue
